feat: use web crypto exclusively, remove dep to node-forge and jws
