+++ apps/website/package.json
@@ -16,10 +16,9 @@
     "build:next-sitemap": "next-sitemap --config next-sitemap.config.cjs",
     "build:push-sw": "tsup --config sw.tsup.config.ts",
     "build:manifest": "tsx ./build-scripts/create-manifest.ts",
-    "build:vapid": "tsx ./build-scripts/vapid.ts",
     "dev:next": "next dev",
     "dev:push-sw": "tsup --watch --config sw.tsup.config.ts",
-    "build": "run-s build:vapid build:next build:next-sitemap build:push-sw build:manifest",
+    "build": "run-s build:next build:next-sitemap build:push-sw build:manifest",
     "dev": "run-p build:vapid build:manifest dev:*",
     "postinstall:prisma": "prisma generate",
     "postinstall:create-data": "tsx ./build-scripts/create-data.ts",
@@ -50,7 +49,6 @@
     "dompurify": "^3.0.2",
     "html-react-parser": "^3.0.16",
     "jsdom": "^21.1.1",
-    "jws": "^4.0.0",
     "lodash": "^4.17.21",
     "luxon": "^3.3.0",
     "next": "^13.3.2",
@@ -58,7 +56,6 @@
     "next-sitemap": "^4.0.7",
     "next-superjson": "^0.0.4",
     "node-fetch": "^3.3.1",
-    "node-forge": "^1.3.1",
     "oauth": "^0.10.0",
     "photoswipe": "^5.3.7",
     "probe-image-size": "^7.2.3",
@@ -87,7 +84,6 @@
     "@types/country-list": "^2.1.1",
     "@types/dompurify": "^2.4.0",
     "@types/jsdom": "^21.1.1",
-    "@types/jws": "^3.2.5",
     "@types/lodash": "^4.14.194",
     "@types/luxon": "^3.3.0",
     "@types/node": "^18.16.3",
