+++ apps/website/src/server/utils/web-push/content-encryption.ts
@@ -1,12 +1,11 @@
-import type { ECDH } from "node:crypto";
-import {
-  createHmac,
-  createCipheriv,
-  createECDH,
-  randomBytes,
-} from "node:crypto";
+import { webcrypto as crypto } from "node:crypto";
 
 import { decodeBase64UrlToArrayBuffer } from "@/utils/base64url";
+import {
+  concatArrayBuffers,
+  createUintArrayBEFromNumber,
+  readNumberFromUintArrayBE,
+} from "@/utils/array-buffer";
 
 export const PAD_SIZE = 1;
 export const TAG_LENGTH = 16;
@@ -15,130 +14,189 @@ export const NONCE_LENGTH = 12;
 export const SHA_256_LENGTH = 32;
 
 type KeyAndNonce = {
-  key: Buffer;
-  nonce: Buffer;
+  key: ArrayBuffer;
+  nonce: ArrayBuffer;
 };
 
-export function HMAC_hash(key: Buffer, input: Buffer) {
-  const hmac = createHmac("sha256", key);
-  hmac.update(input);
-  return hmac.digest();
+export async function createECDH() {
+  return crypto.subtle.generateKey(
+    { name: "ECDH", namedCurve: "P-256" },
+    true,
+    ["deriveKey", "deriveBits"]
+  );
 }
 
-export function HKDF_expand(
-  pseudoRandomKey: Buffer,
-  info: Buffer,
+export async function computeSecret(
+  privateKey: CryptoKey,
+  publicKey: CryptoKey
+) {
+  return crypto.subtle.deriveBits(
+    {
+      name: "ECDH",
+      public: publicKey,
+    },
+    privateKey,
+    256
+  );
+}
+
+export async function HMAC_hash(key: ArrayBuffer, input: ArrayBuffer) {
+  return crypto.subtle.sign(
+    "HMAC",
+    await crypto.subtle.importKey(
+      "raw",
+      key,
+      { name: "HMAC", hash: "SHA-256" },
+      false,
+      ["sign", "verify"]
+    ),
+    input
+  );
+}
+
+export async function HKDF_expand(
+  pseudoRandomKey: ArrayBuffer,
+  info: ArrayBuffer,
   length: number
 ) {
-  const counterBuffer = Buffer.alloc(1);
+  const counterBuffer = new Uint8Array(1);
 
-  let output = Buffer.alloc(0);
-  let T = Buffer.alloc(0);
+  let output = new ArrayBuffer(0);
+  let lastHash = new ArrayBuffer(0);
   let counter = 0;
-  while (output.length < length) {
-    counterBuffer.writeUIntBE(++counter, 0, 1);
-    T = HMAC_hash(pseudoRandomKey, Buffer.concat([T, info, counterBuffer]));
-    output = Buffer.concat([output, T]);
+  while (output.byteLength < length) {
+    counterBuffer[0] = ++counter;
+    const input = concatArrayBuffers([lastHash, info, counterBuffer]);
+    const hash = await HMAC_hash(pseudoRandomKey, input);
+    output = concatArrayBuffers([output, hash]);
+    lastHash = hash;
   }
 
   return output.slice(0, length);
 }
 
-export function HKDF(
-  salt: Buffer,
-  inputKeyMaterial: Buffer,
-  info: Buffer,
+export async function deriveHmacKey(
+  salt: ArrayBuffer,
+  inputKeyMaterial: ArrayBuffer,
+  info: ArrayBuffer,
   length: number
 ) {
-  return HKDF_expand(HMAC_hash(salt, inputKeyMaterial), info, length);
+  const pseudoRandomKey = await HMAC_hash(salt, inputKeyMaterial);
+  return HKDF_expand(pseudoRandomKey, info, length);
 }
 
-export function deriveKeyAndNonce(params: {
-  salt: Buffer;
-  authSecret: Buffer;
-  dh: Buffer;
-  localKeypair: ECDH;
-}): KeyAndNonce {
-  const secret = HKDF(
+export async function deriveKeyAndNonce(params: {
+  salt: ArrayBuffer;
+  authSecret: ArrayBuffer;
+  dh: ArrayBuffer;
+  localKeypair: { privateKey: CryptoKey; publicKey: CryptoKey };
+}): Promise<KeyAndNonce> {
+  const userPublicKey = await crypto.subtle.importKey(
+    "raw",
+    params.dh,
+    { name: "ECDH", namedCurve: "P-256" },
+    true,
+    []
+  );
+
+  const secret = await deriveHmacKey(
     params.authSecret,
-    params.localKeypair.computeSecret(params.dh),
-    Buffer.concat([
-      Buffer.from("WebPush: info\0"),
+    await computeSecret(params.localKeypair.privateKey, userPublicKey),
+    concatArrayBuffers([
+      new TextEncoder().encode("WebPush: info\0"),
       params.dh,
-      params.localKeypair.getPublicKey(),
+      await crypto.subtle.exportKey("raw", params.localKeypair.publicKey),
     ]),
     SHA_256_LENGTH
   );
-  const pseudoRandomKey = HMAC_hash(params.salt, secret);
+  const prk = await HMAC_hash(params.salt, secret);
   return {
-    key: HKDF_expand(
-      pseudoRandomKey,
-      Buffer.from("Content-Encoding: aes128gcm\0"),
+    key: await HKDF_expand(
+      prk,
+      new TextEncoder().encode("Content-Encoding: aes128gcm\0"),
       KEY_LENGTH
     ),
-    nonce: HKDF_expand(
-      pseudoRandomKey,
-      Buffer.from("Content-Encoding: nonce\0"),
+    nonce: await HKDF_expand(
+      prk,
+      new TextEncoder().encode("Content-Encoding: nonce\0"),
       NONCE_LENGTH
     ),
   };
 }
 
-export function generateNonce(base: Buffer, counter: number) {
-  const nonce = Buffer.from(base);
-  const m = nonce.readUIntBE(nonce.length - 6, 6);
+export function generateNonce(base: ArrayBuffer, counter: number) {
+  const nonce = new Uint8Array(base);
+  const m = readNumberFromUintArrayBE(nonce.slice(nonce.length - 6));
   const x =
     ((m ^ counter) & 0xffffff) +
     (((m / 0x1000000) ^ (counter / 0x1000000)) & 0xffffff) * 0x1000000;
-  nonce.writeUIntBE(x, nonce.length - 6, 6);
+
+  nonce.set(createUintArrayBEFromNumber(x, 6), nonce.length - 6);
   return nonce;
 }
 
-export function encryptRecord(
+export async function encryptRecord(
   key: KeyAndNonce,
   counter: number,
-  buffer: Buffer,
+  buffer: ArrayBuffer,
   pad = 0,
   last = false
 ) {
   pad = pad || 0;
   const nonce = generateNonce(key.nonce, counter);
-  const gcm = createCipheriv("aes-128-gcm", key.key, nonce);
 
-  const ciphertext = [];
-  const padding = Buffer.alloc(pad + PAD_SIZE);
+  const aesKey = await crypto.subtle.generateKey(
+    {
+      name: "AES-GCM",
+      length: 256,
+    },
+    true,
+    ["encrypt", "decrypt"]
+  );
+
+  const padding = new Uint8Array(pad + PAD_SIZE);
   padding.fill(0);
+  new DataView(padding.buffer).setUint8(0, last ? 2 : 1);
 
-  ciphertext.push(gcm.update(buffer));
-  padding.writeUIntBE(last ? 2 : 1, 0, 1);
-  ciphertext.push(gcm.update(padding));
+  const ciphertextBuffer = concatArrayBuffers([buffer, padding]);
 
-  gcm.final();
-  const tag = gcm.getAuthTag();
-  if (tag.length !== TAG_LENGTH) {
+  const encrypted = await crypto.subtle.encrypt(
+    {
+      name: "AES-GCM",
+      iv: nonce,
+    },
+    aesKey,
+    buffer
+  );
+  const [, tag] = [
+    encrypted.slice(0, encrypted.byteLength - 16),
+    encrypted.slice(encrypted.byteLength - 16),
+  ];
+  if (tag.byteLength !== TAG_LENGTH) {
     throw new Error("invalid tag generated");
   }
-  ciphertext.push(tag);
-  return Buffer.concat(ciphertext);
+  return concatArrayBuffers([ciphertextBuffer, tag]);
 }
 
-export function writeHeader(keyId: Buffer, rs: number, salt: Buffer) {
-  const ints = Buffer.alloc(5);
-  ints.writeUIntBE(rs, 0, 4);
-  ints.writeUIntBE(keyId.length, 4, 1);
-  return Buffer.concat([salt, ints, keyId]);
+export function writeHeader(keyId: ArrayBuffer, rs: number, salt: Uint8Array) {
+  const ints = new ArrayBuffer(5);
+  const dv = new DataView(ints);
+  dv.setUint32(0, rs, false);
+  dv.setUint8(4, salt.byteLength);
+  return concatArrayBuffers([salt, ints, keyId]);
 }
 
-export function createCipherText(
-  localPublicKey: Buffer,
-  salt: Buffer,
-  payload: Buffer,
+export async function createCipherText(
+  localPublicKey: CryptoKey,
+  salt: Uint8Array,
+  payload: Uint8Array,
   key: KeyAndNonce
 ) {
   const overhead = PAD_SIZE + TAG_LENGTH;
   const rs = 4096;
 
-  let cipherText = writeHeader(localPublicKey, rs, salt);
+  const publicKeyBuffer = await crypto.subtle.exportKey("raw", localPublicKey);
+  let cipherText = writeHeader(publicKeyBuffer, rs, salt);
   let start = 0;
   let pad = 0;
   let counter = 0;
@@ -154,14 +212,14 @@ export function createCipherText(
     const end = start + rs - overhead - recordPad;
     last = end >= payload.length;
     last = last && pad <= 0;
-    const block = encryptRecord(
+    const block = await encryptRecord(
       key,
       counter,
       payload.subarray(start, end),
       recordPad,
       last
     );
-    cipherText = Buffer.concat([cipherText, block]);
+    cipherText = concatArrayBuffers([cipherText, block]);
 
     start = end;
     ++counter;
@@ -170,10 +228,10 @@ export function createCipherText(
   return cipherText;
 }
 
-export function encryptContent(
+export async function encryptContent(
   userPublicKey: string,
   userPrivateKey: string,
-  payload: Buffer
+  payload: Uint8Array
 ) {
   const dh = decodeBase64UrlToArrayBuffer(userPublicKey);
   if (dh.length !== 65) {
@@ -189,15 +247,14 @@ export function encryptContent(
     );
   }
 
-  const localKeypair = createECDH("prime256v1");
-  const localPublicKey = localKeypair.generateKeys();
-  const salt = Buffer.from(randomBytes(16));
-  const key = deriveKeyAndNonce({
+  const localKeypair = await createECDH();
+  const salt = crypto.getRandomValues(new Uint8Array(16));
+  const key = await deriveKeyAndNonce({
     salt,
     localKeypair,
-    dh: Buffer.from(dh),
-    authSecret: Buffer.from(authSecret),
+    dh,
+    authSecret,
   });
 
-  return createCipherText(localPublicKey, salt, payload, key);
+  return createCipherText(localKeypair.publicKey, salt, payload, key);
 }
