+++ apps/website/vitest.config.ts
@@ -8,7 +8,4 @@ export default defineConfig({
       "@": fileURLToPath(new URL("./src", import.meta.url)),
     },
   },
-  test: {
-    watchExclude: [".web-push-vapid.json"],
-  },
 });
